requirements.txt
openai>=1.0.0
google-generativeai
python-pptx
streamlit
